+++ app/dashboard/certificates/components/CertificatesError.tsx
@@ -1,0 +1,8 @@
+export default function CertificatesError({ message }: { message: string }) {
+  return (
+    <div className="text-center text-red-500 py-8">
+      <h2>Error</h2>
+      <p>{message}</p>
+    </div>
+  );
+}
