+++ app/signup/page.tsx
@@ -4,13 +4,7 @@ import { useEffect } from "react";
 import { useRouter } from "next/navigation";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardHeader,
-  CardTitle,
-} from "@/components/ui/card";
+import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card";
 import { useAuth } from "@/lib/auth-context";
 import Link from "next/link";
 import { useForm } from "react-hook-form";
@@ -26,6 +20,7 @@ import {
 } from "@/components/ui/form";
 import { toast } from "sonner";
 import Header from "@/components/header";
+import { signupUser } from "./api/signupUser";
 
 const signupSchema = z
   .object({
@@ -41,9 +36,15 @@ const signupSchema = z
 
 type SignupFormValues = z.infer<typeof signupSchema>;
 
+type SignupResponse = {
+  success: boolean;
+  message?: string;
+  // ...other fields as needed
+};
+
 export default function SignupPage() {
   const router = useRouter();
-  const { signup, isAuthenticated } = useAuth();
+  const { isAuthenticated } = useAuth();
 
   useEffect(() => {
     if (isAuthenticated) {
@@ -64,12 +65,12 @@ export default function SignupPage() {
 
   const onSubmit = async (values: SignupFormValues) => {
     try {
-      const response = await signup(values.email, values.password, values.name);
+      const response: SignupResponse = await signupUser(values.name, values.email, values.password);
       if (response.success) {
         toast.success("Account created successfully. Please login.");
         router.push("/login");
       } else {
-        toast.error(response.message);
+        toast.error(response.message || "Signup failed.");
       }
     } catch (error) {
       console.error(error);
@@ -88,16 +89,11 @@ export default function SignupPage() {
         <Card className="w-[400px]">
           <CardHeader>
             <CardTitle>Create Account</CardTitle>
-            <CardDescription>
-              Enter your details to create a new account
-            </CardDescription>
+            <CardDescription>Enter your details to create a new account</CardDescription>
           </CardHeader>
           <CardContent>
             <Form {...form}>
-              <form
-                onSubmit={form.handleSubmit(onSubmit)}
-                className="space-y-4"
-              >
+              <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-4">
                 <FormField
                   control={form.control}
                   name="name"
@@ -153,13 +149,9 @@ export default function SignupPage() {
                 <Button
                   type="submit"
                   className="w-full cursor-pointer"
-                  disabled={
-                    !form.formState.isValid || form.formState.isSubmitting
-                  }
+                  disabled={!form.formState.isValid || form.formState.isSubmitting}
                 >
-                  {form.formState.isSubmitting
-                    ? "Creating account..."
-                    : "Sign Up"}
+                  {form.formState.isSubmitting ? "Creating account..." : "Sign Up"}
                 </Button>
                 <div className="text-center text-sm">
                   Already have an account?{" "}
