+++ app/student-signup/page.tsx
@@ -6,13 +6,7 @@ import Link from "next/link";
 import { useAuth, BackendAPIResponse } from "@/lib/auth-context";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardHeader,
-  CardTitle,
-} from "@/components/ui/card";
+import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card";
 import {
   Form,
   FormControl,
@@ -26,16 +20,13 @@ import { zodResolver } from "@hookform/resolvers/zod";
 import { z } from "zod";
 import { toast } from "sonner";
 import { makeUrl } from "@/lib/utils";
+import { signupStudent } from "./api/signupStudent";
 
 const studentSignupFormSchema = z.object({
   name: z.string().min(2, { message: "Name must be at least 2 characters." }),
   email: z.string().email({ message: "Please enter a valid email address." }),
-  password: z
-    .string()
-    .min(8, { message: "Password must be at least 8 characters." }),
-  batch_registration_key: z
-    .string()
-    .min(1, { message: "Batch registration key is required." }),
+  password: z.string().min(8, { message: "Password must be at least 8 characters." }),
+  batch_registration_key: z.string().min(1, { message: "Batch registration key is required." }),
 });
 
 type StudentSignupFormValues = z.infer<typeof studentSignupFormSchema>;
@@ -82,57 +73,32 @@ function StudentSignupFormContents() {
   const onSubmit = async (values: StudentSignupFormValues) => {
     setIsLoading(true);
     try {
-      const response = await fetch(makeUrl("signupStudent"), {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify(values),
-        credentials: "include",
-      });
-
-      const result = await response.json(); // result can be StudentSignupApiResponse or HTTPValidationError
-
-      if (!response.ok) {
-        // Handle HTTP errors (e.g., 422 Validation Error)
-        let httpErrorMessage = "Student signup failed. Please try again.";
-        const validationError = result as HTTPValidationError;
-        if (validationError.detail && Array.isArray(validationError.detail)) {
-          httpErrorMessage = validationError.detail
-            .map((err) => `${err.loc.join(".") || "field"} - ${err.msg}`)
-            .join(", ");
-        } else if ((result as StudentSignupApiResponse).message) {
-          httpErrorMessage = (result as StudentSignupApiResponse).message!;
-        } else if (typeof result.error === "string") {
-          httpErrorMessage = result.error;
-        }
-        toast.error(httpErrorMessage);
-      } else if (!result.success) {
-        // Handle cases where response is OK (200) but operation failed (result.success === false)
-        let apiErrorMessage =
-          result.message || "Student signup failed. Please try again.";
-        if (result.error) {
-          if (typeof result.error === "string") {
-            apiErrorMessage = result.error;
-          } else if (Array.isArray(result.error)) {
-            apiErrorMessage = result.error.join(", ");
+      const response = await signupStudent(
+        values.name,
+        values.email,
+        values.password,
+        values.batch_registration_key,
+      );
+      if (!response.success) {
+        let apiErrorMessage = response.message || "Student signup failed. Please try again.";
+        if (response.error) {
+          if (typeof response.error === "string") {
+            apiErrorMessage = response.error;
+          } else if (Array.isArray(response.error)) {
+            apiErrorMessage = response.error.join(", ");
           } else {
-            apiErrorMessage = JSON.stringify(result.error);
+            apiErrorMessage = JSON.stringify(response.error);
           }
         }
         toast.error(apiErrorMessage);
       } else {
-        // Successful signup
-        toast.success(
-          result.message || "Signup successful! Redirecting to dashboard..."
-        );
+        toast.success(response.message || "Signup successful! Redirecting to dashboard...");
         await fetchUserOnMount();
         router.push("/dashboard?signup=success");
       }
     } catch (error) {
       console.error("Student signup error:", error);
-      const message =
-        error instanceof Error
-          ? error.message
-          : "An unexpected error occurred.";
+      const message = error instanceof Error ? error.message : "An unexpected error occurred.";
       toast.error(`Student signup failed: ${message}`);
     } finally {
       setIsLoading(false);
@@ -158,12 +124,9 @@ function StudentSignupFormContents() {
     <div className="flex min-h-screen items-center justify-center bg-gray-100 dark:bg-gray-900 p-4">
       <Card className="w-full max-w-md">
         <CardHeader className="text-center">
-          <CardTitle className="text-2xl font-bold">
-            Student Registration
-          </CardTitle>
+          <CardTitle className="text-2xl font-bold">Student Registration</CardTitle>
           <CardDescription>
-            Create your account using the registration key provided by your
-            instructor.
+            Create your account using the registration key provided by your instructor.
           </CardDescription>
         </CardHeader>
         <CardContent>
@@ -189,11 +152,7 @@ function StudentSignupFormContents() {
                   <FormItem>
                     <FormLabel>Email</FormLabel>
                     <FormControl>
-                      <Input
-                        type="email"
-                        placeholder="your.email@example.com"
-                        {...field}
-                      />
+                      <Input type="email" placeholder="your.email@example.com" {...field} />
                     </FormControl>
                     <FormMessage />
                   </FormItem>
@@ -206,11 +165,7 @@ function StudentSignupFormContents() {
                   <FormItem>
                     <FormLabel>Password</FormLabel>
                     <FormControl>
-                      <Input
-                        type="password"
-                        placeholder="********"
-                        {...field}
-                      />
+                      <Input type="password" placeholder="********" {...field} />
                     </FormControl>
                     <FormMessage />
                   </FormItem>
@@ -236,10 +191,7 @@ function StudentSignupFormContents() {
           </Form>
           <p className="mt-6 text-center text-sm text-gray-600 dark:text-gray-400">
             Already have an account?{" "}
-            <Link
-              href="/login"
-              className="font-medium text-primary hover:underline"
-            >
+            <Link href="/login" className="font-medium text-primary hover:underline">
               Log in here
             </Link>
           </p>
