+++ docs/refactor-report.md
@@ -1,0 +1,325 @@
+# Next.js Application Refactor Report
+
+## 1. Overview
+
+### Purpose of the Refactor
+
+- Modernize the codebase to leverage Next.js App Router best practices (v15+)
+- Improve maintainability, scalability, and developer experience
+- Optimize performance and bundle size
+- Enforce strict typing, linting, and testing standards
+
+### High-Level Summary of Improvements
+
+- Adopted feature-based folder structure
+- Clear separation of client and server components
+- Centralized and optimized data fetching
+- Modularized and isolated components
+- Enhanced performance via lazy loading and dynamic imports
+- Enforced strict TypeScript and linting
+- Improved onboarding and documentation
+
+---
+
+## 2. Folder Structure Before & After
+
+### Original Project Structure (Outline)
+
+```plaintext
+app/
+  admin/
+  dashboard/
+  login/
+  signup/
+  student-signup/
+  callback/
+  ...
+components/
+  ui/
+  ...
+lib/
+  ...
+public/
+  ...
+```
+
+- Components and utilities were globally grouped, with some feature folders under `app/`.
+
+### New Proposed Structure
+
+```plaintext
+app/
+  dashboard/
+    components/
+    api/
+    hooks/
+    ...
+  admin/
+    components/
+    api/
+    hooks/
+    ...
+  login/
+    api/
+    ...
+  signup/
+    api/
+    ...
+  student-signup/
+    api/
+    ...
+  callback/
+    api/
+    ...
+components/
+  ui/
+  shared/
+lib/
+  ...
+docs/
+  refactor-report.md
+```
+
+- **Feature-based**: Each major route/feature has its own folder with `components/`, `api/`, and optionally `hooks/`.
+- **UI components**: Atomic, reusable UI in `components/ui/`.
+- **Shared components**: Cross-feature components in `components/shared/`.
+- **Docs**: All documentation in `/docs`.
+
+---
+
+## 3. Client vs Server Component Refactor
+
+### What Changed
+
+- Defaulted to **server components** for all pages/layouts unless interactivity is required.
+- **Client components** only for UI with state, effects, or event handlers.
+- `"use client"` is now only present in files that require React hooks, event handlers, or browser APIs.
+
+#### Example:
+
+```tsx
+// Server component (default)
+export default async function DashboardPage() {
+  // ...fetch data on the server
+}
+
+// Client component (interactivity)
+("use client");
+export default function DashboardStats({ data }) {
+  // ...uses useState, useEffect, etc.
+}
+```
+
+### Why the Change Was Made
+
+- Reduces client bundle size
+- Improves security by keeping sensitive logic server-side
+- Follows Next.js best practices for performance and maintainability
+
+---
+
+## 4. Data Fetching Strategy Refactor
+
+### Previous Issues
+
+- Data fetching logic was often duplicated or scattered
+- Fetches sometimes occurred in client components when they could be server-side
+- Poor separation of concerns
+
+### New Strategy
+
+- **Centralized fetch logic** in `api/` folders within each feature
+- **Server components** handle all SSR/SSG and sensitive fetches
+- **Client components** only fetch when absolutely necessary (e.g., after user interaction)
+
+#### Example Fetch Function
+
+```ts
+// app/dashboard/api/fetchDashboardData.ts
+"use server";
+import { cookies } from "next/headers";
+import { makeUrl } from "@/lib/utils";
+export async function fetchDashboardData(email: string) {
+  // ...fetch logic
+}
+```
+
+#### Usage Example
+
+```tsx
+// app/dashboard/page.tsx
+import { fetchDashboardData } from "./api/fetchDashboardData";
+export default async function DashboardPage() {
+  const data = await fetchDashboardData(user.email);
+  // ...
+}
+```
+
+---
+
+## 5. Codebase Modularity & Component Isolation
+
+### Improvements
+
+- Extracted reusable UI elements to `components/ui/`
+- Feature-specific components moved to their respective `components/` folders
+- Applied atomic design principles for UI (atoms, molecules, etc.)
+- Reduced code duplication and improved DRYness
+
+#### Example
+
+- `DashboardStats`, `CertificatesList`, `DemosList` are now isolated, reusable, and colocated with their feature
+
+---
+
+## 6. Performance Optimizations
+
+### Changes Made
+
+- Removed unnecessary `"use client"` directives
+- Defaulted to server components for all pages/layouts
+- Used dynamic imports for heavy or infrequently used client components
+- Prepared for React Suspense boundaries (where supported)
+
+#### Example
+
+```tsx
+import dynamic from "next/dynamic";
+const HeavyChart = dynamic(() => import("./HeavyChart"), { ssr: false });
+```
+
+### Metrics
+
+- Bundle size reduction and improved TTFB (if measured)
+
+---
+
+## 7. Linting, Type Safety & Testing
+
+### Linting & Formatting
+
+- Updated ESLint config to extend Next.js, TypeScript, and Prettier
+- Added `.prettierrc` for consistent formatting
+
+### TypeScript
+
+- Enforced `strict: true` in `tsconfig.json`
+- Improved types for API responses and form values
+
+#### Example: Before/After
+
+**Before:**
+
+```ts
+const [data, setData] = useState<any>(null);
+```
+
+**After:**
+
+```ts
+import type { DashboardData } from "@/lib/dashboard-types";
+const [data, setData] = useState<DashboardData | null>(null);
+```
+
+### Testing
+
+- Recommended: Use React Testing Library and Jest for unit/integration tests
+- Example test files added for key components
+
+---
+
+## 8. Deployment Readiness & DevOps
+
+### Dockerfile
+
+- Ensure multi-stage builds for smaller images
+- Use `next start` for production
+
+### next.config.js
+
+- No breaking changes; ensure best practices for environment variables and security
+
+### .env Handling
+
+- Use `NEXT_PUBLIC_` prefix for client-exposed env vars
+- Keep secrets server-side only
+
+---
+
+## 9. Developer Experience Enhancements
+
+### README Improvements
+
+- Added onboarding, project structure, and conventions
+- Documented scripts and development workflow
+
+### Tooling
+
+- Husky and code generators (e.g., Plop) recommended but optional
+- Prettier and ESLint enforced
+
+### Project Conventions
+
+- Feature-based folders
+- Centralized fetch logic
+- Strict typing and linting
+
+---
+
+## 10. Before & After Comparison
+
+### Example: Dashboard Data Fetching
+
+**Before:**
+
+```tsx
+// app/dashboard/page.tsx
+"use client";
+useEffect(() => { fetch(...) }, []);
+```
+
+**After:**
+
+```tsx
+// app/dashboard/page.tsx
+export default async function DashboardPage() {
+  const data = await fetchDashboardData(...);
+  return <DashboardStats data={data} />;
+}
+```
+
+### Example: Component Placement
+
+**Before:**
+
+```
+components/DashboardStats.tsx
+```
+
+**After:**
+
+```
+app/dashboard/components/DashboardStats.tsx
+```
+
+---
+
+## 11. Migration Guide
+
+### For Developers Migrating from the Old Structure
+
+- **Find feature code** under `app/[feature]/` (e.g., `app/dashboard/`)
+- **Look for fetch logic** in `api/` folders within each feature
+- **UI components** are colocated with their feature, or in `components/ui/` if shared
+- **Use server components** by default; only add `"use client"` for interactivity
+
+### Contributing to the New Structure
+
+1. Add new features as folders under `app/`
+2. Place server fetch logic in `api/`, UI in `components/`, and hooks in `hooks/`
+3. Use strict types and follow linting/formatting rules
+4. Update documentation in `/docs` as needed
+
+---
+
+**This document should be maintained as the canonical reference for architecture, onboarding, and future refactors.**
